More demo fixes [skip ci]

+++ demo/undocumented/collision-detection/python/demo_collision-detection.py
@@ -40,7 +40,7 @@ mesh_B.translate(Point(x_B, y_B))
 mesh_C.translate(Point(x_C, y_C))
 
 # Create mesh function for plotting
-f = CellFunction("uint", mesh_A)
+f = CellFunction("size_t", mesh_A)
 
 # Build bounding box trees for background mesh
 tree_A = BoundingBoxTree()
